server/requirements.txt
asgiref==3.11.0
channels==4.3.2
channels_redis==4.3.0
python-decouple==3.8
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.4
tzdata==2025.3
